+++ Vault-Coder/Changes.md
@@ -8,6 +8,8 @@ Parent module updates
 ### New features
 - Added two new data types: **Vector2D** and **Distance** (from **Paradigm**) and **LatLong** (from **Terra**)
 - You may now specify an extending upper length limit within the type declaration, e.g. by writing `"String(3-9)"`
+- Added default database property (i.e. part) names for the pre-defined multi-column data types, 
+  as well as support for default part names in custom data types
 ### Other changes
 - Setting `"crop": true` to a property without a length limit now sets the column length rule to `"crop"`